Re-Added to_numpy()
